# ref: grid cleanup and decoupling the bin storage