--- a/core/include/detray/coordinates/cartesian2.hpp
+++ b/core/include/detray/coordinates/cartesian2.hpp
@@ -57,12 +57,21 @@ struct cartesian2 final : public coordinate_base<cartesian2, transform3_t> {
 
     /// @}
 
-    /** This method transform from a point from global cartesian 3D frame to a
-     * local 2D cartesian point */
+    /** This method transforms a point from a global cartesian 3D frame to a
+     * local 3D cartesian point */
     DETRAY_HOST_DEVICE
     inline point3 global_to_local(const transform3_t &trf3, const point3 &p,
                                   const vector3 & /*d*/) const {
         return trf3.point_to_local(p);
+    }
+
+    /** This method transforms a point from a global cartesian 3D frame to a
+     * bound 2D cartesian point */
+    DETRAY_HOST_DEVICE
+    inline loc_point project_to_axes(const transform3_t &trf3, const point3 &p,
+                                     const vector3 & /*d*/) const {
+        auto loc_p = trf3.point_to_local(p);
+        return {loc_p[0], loc_p[1]};
     }
 
     /** This method transform from a local 2D cartesian point to a point global