--- a/core/include/detray/materials/material_map.hpp
+++ b/core/include/detray/materials/material_map.hpp
@@ -10,7 +10,9 @@
 // Detray include(s)
 #include "detray/definitions/containers.hpp"
 #include "detray/materials/material_slab.hpp"
+#include "detray/surface_finders/grid/detail/grid_bins.hpp"
 #include "detray/surface_finders/grid/grid.hpp"
+#include "detray/surface_finders/grid/serializers.hpp"
 #include "detray/tools/grid_factory.hpp"
 
 namespace detray {
@@ -18,13 +20,14 @@ namespace detray {
 /// Definition of binned material
 template <typename axes_shape, typename scalar_t,
           typename container_t = host_container_types, bool owning = false>
-using material_map = grid<coordinate_axes<axes_shape, owning, container_t>,
-                          material_slab<scalar_t>, simple_serializer, replacer>;
+using material_map =
+    grid<coordinate_axes<axes_shape, owning, container_t>,
+         bins::single<material_slab<scalar_t>>, simple_serializer>;
 
 /// How to build material maps of various shapes
 // TODO: Move to material_map_builder once available
 template <typename scalar_t = detray::scalar>
 using material_map_factory =
-    grid_factory<material_slab<scalar_t>, simple_serializer, replacer>;
+    grid_factory<bins::single<material_slab<scalar_t>>, simple_serializer>;
 
 }  // namespace detray