--- a/core/include/detray/surface_finders/grid/serializers.hpp
+++ b/core/include/detray/surface_finders/grid/serializers.hpp
@@ -1,6 +1,6 @@
 /** Detray library, part of the ACTS project (R&D line)
  *
- * (c) 2022 CERN for the benefit of the ACTS project
+ * (c) 2022-2024 CERN for the benefit of the ACTS project
  *
  * Mozilla Public License Version 2.0
  */
@@ -8,4 +8,4 @@
 #pragma once
 
 // Project include(s).
-#include "detray/surface_finders/grid/detail/serializer_impl.hpp"
+#include "detray/surface_finders/grid/detail/simple_serializer.hpp"