This PR performs some straight-forward cleanups in the grid and axis types (mostly renaming) and decouples the populator implementation from the bin storage. Previously, the type of populator would determine the type of bin in the grid storage, which made it very difficult to define a grid with runtime bin capacities. Now, the populator type is only griven on the populate method of the grid and the bin storage and bin types are implemented separately:
- ```single```: A bin with a single entry
- ```static_array```: A bin with an array of entries (capacity is defined at compile-time)
- ```dynamic_array```: A bin with runtime bin capacity (holds an index pair into the bin entry storage)