@niermann999 2024-01-13
This is not yet implementing anything new, just setting the stage, but it seems to have had some positive impact on performance, too:

Current MAIN:
```
2024-01-13T14:34:09+01:00
Running ./bin/detray_array_propagator_cuda
Run on (48 X 2370.44 MHz CPU s)
CPU Caches:
  L1 Data 32 KiB (x24)
  L1 Instruction 32 KiB (x24)
  L2 Unified 512 KiB (x24)
  L3 Unified 32768 KiB (x4)
Load Average: 6.97, 7.04, 4.55
--------------------------------------------------------------------------------------
Benchmark                            Time             CPU   Iterations UserCounters...
--------------------------------------------------------------------------------------
CPU unsync propagation/8         60703 ns        60753 ns        11986 TracksPropagated=1053.45k/s
CPU unsync propagation/16       187594 ns       187333 ns         3758 TracksPropagated=1.36655M/s
CPU unsync propagation/32       894820 ns       883139 ns          838 TracksPropagated=1.1595M/s
CPU unsync propagation/64      3466768 ns      3458045 ns          202 TracksPropagated=1.18448M/s
CPU unsync propagation/128    14134561 ns     14098821 ns           50 TracksPropagated=1.16208M/s
CPU unsync propagation/256    55186135 ns     53497915 ns           13 TracksPropagated=1.22502M/s
CPU sync propagation/8           62004 ns        61977 ns        11399 TracksPropagated=1032.64k/s
CPU sync propagation/16         189271 ns       189100 ns         3724 TracksPropagated=1.35378M/s
CPU sync propagation/32         841396 ns       839668 ns          833 TracksPropagated=1.21953M/s
CPU sync propagation/64        3490113 ns      3482087 ns          201 TracksPropagated=1.17631M/s
CPU sync propagation/128      14173918 ns     14140280 ns           49 TracksPropagated=1.15868M/s
CPU sync propagation/256      55269698 ns     52721329 ns           13 TracksPropagated=1.24306M/s
CUDA unsync propagation/8      1429823 ns      1414853 ns          473 TracksPropagated=45.2344k/s
CUDA unsync propagation/16     2033710 ns      2022504 ns          348 TracksPropagated=126.576k/s
CUDA unsync propagation/32     2693060 ns      2674286 ns          253 TracksPropagated=382.906k/s
CUDA unsync propagation/64     3541619 ns      3515899 ns          197 TracksPropagated=1.16499M/s
CUDA unsync propagation/128    5474129 ns      5425841 ns          129 TracksPropagated=3.01962M/s
CUDA unsync propagation/256   19593028 ns     19424051 ns           36 TracksPropagated=3.37396M/s
CUDA sync propagation/8        1345813 ns      1333405 ns          524 TracksPropagated=47.9974k/s
CUDA sync propagation/16       1985537 ns      1974792 ns          360 TracksPropagated=129.634k/s
CUDA sync propagation/32       2538111 ns      2518404 ns          277 TracksPropagated=406.607k/s
CUDA sync propagation/64       3359094 ns      3333127 ns          208 TracksPropagated=1.22888M/s
CUDA sync propagation/128      5214221 ns      5167407 ns          136 TracksPropagated=3.17064M/s
CUDA sync propagation/256     17904616 ns     17732784 ns           40 TracksPropagated=3.69575M/s
```
This PR:
```
Run on (48 X 2369.83 MHz CPU s)
CPU Caches:
  L1 Data 32 KiB (x24)
  L1 Instruction 32 KiB (x24)
  L2 Unified 512 KiB (x24)
  L3 Unified 32768 KiB (x4)
Load Average: 8.00, 5.10, 2.04
--------------------------------------------------------------------------------------
Benchmark                            Time             CPU   Iterations UserCounters...
--------------------------------------------------------------------------------------
CPU unsync propagation/8         58695 ns        58666 ns        12436 TracksPropagated=1090.92k/s
CPU unsync propagation/16       179623 ns       179355 ns         3889 TracksPropagated=1.42734M/s
CPU unsync propagation/32       806516 ns       804718 ns          870 TracksPropagated=1.2725M/s
CPU unsync propagation/64      3343427 ns      3336214 ns          209 TracksPropagated=1.22774M/s
CPU unsync propagation/128    13637111 ns     13605868 ns           51 TracksPropagated=1.20419M/s
CPU unsync propagation/256    52291525 ns     50813815 ns           13 TracksPropagated=1.28973M/s
CPU sync propagation/8           97823 ns        97309 ns        11349 TracksPropagated=657.7k/s
CPU sync propagation/16         180565 ns       180340 ns         3907 TracksPropagated=1.41954M/s
CPU sync propagation/32         813748 ns       811796 ns          864 TracksPropagated=1.2614M/s
CPU sync propagation/64        3348537 ns      3340846 ns          210 TracksPropagated=1.22604M/s
CPU sync propagation/128      13658280 ns     13626228 ns           51 TracksPropagated=1.20239M/s
CPU sync propagation/256      52859884 ns     50920976 ns           13 TracksPropagated=1.28701M/s
CUDA unsync propagation/8      1314907 ns      1303178 ns          493 TracksPropagated=49.1107k/s
CUDA unsync propagation/16     1965386 ns      1950595 ns          366 TracksPropagated=131.242k/s
CUDA unsync propagation/32     2572105 ns      2554568 ns          271 TracksPropagated=400.851k/s
CUDA unsync propagation/64     3391314 ns      3366439 ns          207 TracksPropagated=1.21672M/s
CUDA unsync propagation/128    5343430 ns      5291280 ns          132 TracksPropagated=3.09642M/s
CUDA unsync propagation/256   16830825 ns     16667588 ns           42 TracksPropagated=3.93194M/s
CUDA sync propagation/8        1280218 ns      1267779 ns          550 TracksPropagated=50.482k/s
CUDA sync propagation/16       1831358 ns      1820693 ns          391 TracksPropagated=140.606k/s
CUDA sync propagation/32       2406416 ns      2387793 ns          293 TracksPropagated=428.848k/s
CUDA sync propagation/64       3207411 ns      3181347 ns          217 TracksPropagated=1.28751M/s
CUDA sync propagation/128      5044837 ns      4991901 ns          139 TracksPropagated=3.28212M/s
CUDA sync propagation/256     15700173 ns     15540237 ns           45 TracksPropagated=4.21718M/s
```